requirements.txt
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dnspython==2.7.0
idna==3.10
paramiko==3.5.1
phonenumbers==9.0.8
pillow==11.2.1
pycparser==2.22
PyNaCl==1.5.0
python-nmap==0.7.1
qrcode==8.2
requests==2.32.4
soupsieve==2.7
Sublist3r==1.0
typing_extensions==4.14.0
urllib3==2.5.0
